Fix tests for IE11
