+++ can-type-ie-test.js
@@ -1,0 +1,1 @@
+require('./can-type-test');
