+++ can-type-test.js
@@ -3,7 +3,8 @@ var canReflect = require("can-reflect");
 var type = require("../can-type");
 var QUnit = require("steal-qunit");
 var dev = require("can-test-helpers").dev;
-var helpers = require("./helpers");
+
+require('./type-errors-test');
 
 var newSymbol = canSymbol.for("can.new");
 var isMemberSymbol = canSymbol.for("can.isMember");
@@ -408,25 +409,3 @@ QUnit.test("type.convertAll is a convenience for type.all(type.convert, Type)", 
 	assert.equal(person.last, "Phillips");
 	assert.equal(person.age, 8);
 });
-
-var testCases = [
-	{expectedTypeName: 'Number', value: '3', ExpectedType: Number},
-	{expectedTypeName: 'String', value: 3, ExpectedType: String},
-	{expectedTypeName: 'Number', value: false, ExpectedType: Number},
-	{expectedTypeName: 'Boolean', value: '', ExpectedType: Boolean}
-];
-
-testCases.forEach(function(testCase) {
-	var typeName = helpers.capitalizeTypeName(typeof testCase.value);
-	var expectedTypeName = testCase.expectedTypeName;
-	QUnit.test('Include the type ' + typeName + ' of the value', function(assert) {
-		var strictType = type.check(testCase.ExpectedType);
-	
-		try {
-			canReflect.convert(testCase.value, strictType);
-		} catch (error) {
-			assert.equal(error.message, testCase.value + ' (' + typeName + ') is not of type ' + expectedTypeName + '.');
-		}
-	});
-});
-
