+++ test-ie.html
@@ -1,4 +1,4 @@
 <!doctype html>
 <title>can-type tests</title>
-<script src="node_modules/steal/steal-with-promises.js" main="~/can-type-test"></script>
+<script src="node_modules/steal/steal-with-promises.js" main="~/can-type-ie-test"></script>
 <div id="qunit-fixture"></div>
